public/requirements.txt
future==0.18.2
joblib==0.16.0
numpy==1.19.2
pandas==1.1.2
Pillow==7.2.0
protobuf==3.13.0
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
tensorboardX==2.1
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
